エネミーの処理追加
エネミーの攻撃力を返す関数を追加
プレイヤーの攻撃で消えるように
