--- ProjectTeamf6/Assets/Hosoda.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9600a39de30369a468a6b493d1df6146
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
