--- ProjectTeamf6/Assets/Koba/Material/pb_Enemy.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 85f1352945be78b4d82b1a884d0ad513
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
