+++ ProjectTeamf6/Assets/Koba/Prefab.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 7eeb8c16dced3bd40abf61e70920fa29
-PrefabImporter:
+guid: 2311ce89f455ba94393753ce654aa2d9
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
