+++ ProjectTeamf6/Assets/Koba/Prefab/player_Enemy.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &1403758450330364927
+--- !u!1 &3942048175537981177
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,38 +8,39 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 4823931523967846556}
-  - component: {fileID: 6504055438868749342}
-  - component: {fileID: 2294613789203145421}
-  - component: {fileID: 5882791000354636533}
+  - component: {fileID: 1885820115894178637}
+  - component: {fileID: 8588949297128574768}
+  - component: {fileID: 7882536909706371476}
+  - component: {fileID: 5171670333301279033}
+  - component: {fileID: 198384213472687989}
   m_Layer: 0
-  m_Name: pb_Enemy
+  m_Name: player_Enemy
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &4823931523967846556
+--- !u!4 &1885820115894178637
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1403758450330364927}
+  m_GameObject: {fileID: 3942048175537981177}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 5.79, y: 1.12, z: 0}
-  m_LocalScale: {x: 0.5, y: 0.5, z: 0.5}
+  m_LocalPosition: {x: 0.91, y: -3.02, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!212 &6504055438868749342
+--- !u!212 &8588949297128574768
 SpriteRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1403758450330364927}
+  m_GameObject: {fileID: 3942048175537981177}
   m_Enabled: 1
   m_CastShadows: 0
   m_ReceiveShadows: 0
@@ -73,7 +74,7 @@ SpriteRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_Sprite: {fileID: 21300000, guid: dee8da1573e49a647b2def9788f6355b, type: 3}
-  m_Color: {r: 0, g: 0.47372842, b: 1, a: 1}
+  m_Color: {r: 0.5073457, g: 0, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
@@ -83,36 +84,31 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!114 &2294613789203145421
+--- !u!114 &7882536909706371476
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1403758450330364927}
+  m_GameObject: {fileID: 3942048175537981177}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9a948df2c6fce634288ef18e11fc863c, type: 3}
+  m_Script: {fileID: 11500000, guid: 15d9a2283f93ad847814dc734ff74766, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   health: 5
-  damege: 5
-  speed: 5
-  updateTime: 3
+  damage: 5
+  speed: 0.1
+  playerScript: {fileID: 0}
   player: {fileID: 0}
-  playerBase: {fileID: 0}
-  village1: {fileID: 0}
-  village2: {fileID: 0}
-  village3: {fileID: 0}
-  village4: {fileID: 0}
---- !u!50 &5882791000354636533
+--- !u!50 &5171670333301279033
 Rigidbody2D:
   serializedVersion: 4
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1403758450330364927}
+  m_GameObject: {fileID: 3942048175537981177}
   m_BodyType: 0
   m_Simulated: 1
   m_UseFullKinematicContacts: 0
@@ -125,4 +121,30 @@ Rigidbody2D:
   m_Interpolate: 0
   m_SleepingMode: 1
   m_CollisionDetection: 0
-  m_Constraints: 0
+  m_Constraints: 4
+--- !u!61 &198384213472687989
+BoxCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3942048175537981177}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0, y: 0}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 2, y: 2}
+    newSize: {x: 2, y: 2}
+    adaptiveTilingThreshold: 0.5
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  serializedVersion: 2
+  m_Size: {x: 2, y: 2}
+  m_EdgeRadius: 0
