--- ProjectTeamf6/Assets/Koba/Script/EnemyMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 079cab2d8594c8147b1dc4142ddf73c6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
