--- ProjectTeamf6/Assets/Koba/Script/Pb_EnemyMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e3bd70b9ca3bab142a41bb6f146eca36
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
