--- ProjectTeamf6/Assets/Shimizu.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f3b009015096285459d32771d7bbb510
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
